requirements/local.txt
beautifulsoup4==4.8.2
certifi==2019.11.28
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
cycler==0.10.0
demjson==2.2.4
Django==2.2.5
djangorestframework==3.11.0
drf-yasg==1.17.1
idna==2.9
inflection==0.3.1
itypes==1.1.0
Jinja2==2.11.1
jtutils==0.0.6
kiwisolver==1.1.0
leven==1.0.4
MarkupSafe==1.1.1
nose==1.3.7
packaging==20.3
postgres==3.0.0
psycopg2-binary==2.8.4
psycopg2-pool==1.1
pyparsing==2.4.6
python-csv==0.0.11
python-dateutil==2.8.1
python-decouple==3.1
pytz==2019.3
requests==2.23.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.14.0
soupsieve==2.0
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.8
xlrd==1.2.0
xmltodict==0.12.0
